requirements.txt
Django==5.2.4
# psycopg2-binary==2.9.7
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1
python-dotenv==1.0.0
whitenoise==6.5.0
django-redis==5.3.0
django-ratelimit==4.1.0
django-cors-headers==4.3.1
Pillow==10.0.0
requests==2.31.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
markdown2==2.4.10
